Update to the last rust alpha.
Version of rustc: 1.0.0-nightly (44a287e6e 2015-01-08 17:03:40 -0800).

+++ examples/recording.rs
@@ -17,8 +17,8 @@ fn main() {
     window.set_light(Light::StickToCamera);
 
     let mut recorder = Recorder::new(Path::new("test.mpg"),
-    window.width()  as uint,
-    window.height() as uint);
+    window.width()  as usize,
+    window.height() as usize);
 
     while window.render() {
         c.prepend_to_local_rotation(&Vec3::new(0.0f32, 0.014, 0.0));
