+++ src/builtin/normals_material.rs
@@ -43,7 +43,7 @@ impl NormalsMaterial {
 
 impl Material for NormalsMaterial {
     fn render(&mut self,
-              pass:      uint,
+              pass:      usize,
               transform: &Iso3<f32>,
               scale:     &Vec3<f32>,
               camera:    &mut Camera,
