+++ src/builtin/uvs_material.rs
@@ -43,7 +43,7 @@ impl UvsMaterial {
 
 impl Material for UvsMaterial {
     fn render(&mut self,
-              pass:      uint,
+              pass:      usize,
               transform: &Iso3<f32>,
               scale:     &Vec3<f32>,
               camera:    &mut Camera,
